+++ src/Farmer/Arm/ResourceGroup.fs
@@ -34,7 +34,8 @@ type ParameterValue =
 
 /// Represents all configuration information to generate an ARM template.
 type ResourceGroupDeployment =
-    { Name: ResourceName
+    { TargetResourceGroup: ResourceName
+      DeploymentName: ResourceName
       Dependencies: ResourceId Set
       Outputs : Map<string, string>
       Location : Location
@@ -44,7 +45,7 @@ type ResourceGroupDeployment =
       SubscriptionId : System.Guid option
       Mode: DeploymentMode
       Tags: Map<string,string> }
-    member this.ResourceId = resourceGroupDeployment.resourceId this.Name
+    member this.ResourceId = resourceGroupDeployment.resourceId this.DeploymentName
     /// Secure parameters for resources defined in this template.
     member this.Parameters =
           [ for resource in this.Resources do
@@ -59,7 +60,7 @@ type ResourceGroupDeployment =
                 (function 
                 | :? ResourceGroupDeployment as rg -> rg.RequiredResourceGroups 
                 | _ ->  [])
-        List.distinct (this.Name.Value :: nestedRgs)        
+        List.distinct (this.TargetResourceGroup.Value :: nestedRgs)        
     member this.Template = 
         { Parameters = this.Parameters
           Outputs = this.Outputs |> Map.toList
@@ -71,11 +72,11 @@ type ResourceGroupDeployment =
             let inputParameterKeys = this.ParameterValues |> Seq.map (fun p -> p.Key) |> Set
             this.Parameters |> List.where(fun p -> inputParameterKeys |> Set.contains p.Key |> not)
     interface IArmResource with
-        member this.ResourceId = resourceGroupDeployment.resourceId this.Name
+        member this.ResourceId = this.ResourceId
         member this.JsonModel = 
-            {| resourceGroupDeployment.Create(this.Name, this.Location, dependsOn = this.Dependencies, tags = this.Tags ) with
+            {| resourceGroupDeployment.Create(this.DeploymentName, this.Location, dependsOn = this.Dependencies, tags = this.Tags ) with
                 location = null // location is not supported for nested resource groups
-                resourceGroup = this.Name.Value
+                resourceGroup = this.TargetResourceGroup.Value
                 subscriptionId = this.SubscriptionId |> Option.map string<System.Guid> |> Option.toObj
                 properties = 
                     {|  template = TemplateGeneration.processTemplate this.Template